bot/requirements.txt
botbuilder-integration-aiohttp>=4.14.0
requests>=2.32.3
botbuilder-dialogs>=4.15.0